+++ packages/common/web/src/Base/createBaseComponent.jsx
@@ -2,7 +2,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { findDOMNode } from 'react-dom';
-import { compact, isNil } from 'lodash';
+import { merge, compact, isNil } from 'lodash';
 import classNames from 'classnames';
 import createHocName from '../Utils/createHocName';
 import getDisplayName from '../Utils/getDisplayName';
@@ -51,17 +51,26 @@ const createBaseComponent = ({
     }
 
     stateFromProps(props) {
-      const { getData, readOnly } = props.blockProps;
+      const { readOnly } = props.blockProps;
       const initialState = pubsub.get('initialState_' + props.block.getKey());
       if (initialState) {
         //reset the initial state
         pubsub.set('initialState_' + props.block.getKey(), undefined);
       }
       return {
-        componentData: getData() || { config: DEFAULTS },
+        componentData: this.getData(props),
         readOnly: !!readOnly,
         componentState: initialState || {},
       };
+    }
+
+    getData(props) {
+      const { getData } = props.blockProps;
+      const data = getData() || { config: DEFAULTS };
+      if (settings?.defaultData) {
+        merge(data, settings.defaultData);
+      }
+      return data;
     }
 
     componentDidMount() {